Make -identify's 'ID_LENGTH=' print a float and not an integer.. The accuracey may be totally fake for some demuxers (mpg), but accurate for others.. (avi)
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@16347 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/demux_asf.c
@@ -422,7 +422,7 @@ static int demux_asf_control(demuxer_t *demuxer,int cmd, void *arg){
 */
     switch(cmd) {
 	case DEMUXER_CTRL_GET_TIME_LENGTH:
-	    *((unsigned long *)arg)=(unsigned long)(asf_movielength);
+	    *((double *)arg)=(double)(asf_movielength);
 	    return DEMUXER_CTRL_OK;
 
 	case DEMUXER_CTRL_GET_PERCENT_POS:
